+++ examples/discovery-search-app/CHANGELOG.md
@@ -3,6 +3,49 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [1.1.0-beta.1](https://github.com/watson-developer-cloud/discovery-components/compare/discovery-search-app@1.1.0-beta.0...discovery-search-app@1.1.0-beta.1) (2020-03-09)
+
+**Note:** Version bump only for package discovery-search-app
+
+
+
+
+
+# [1.1.0-beta.0](https://github.com/watson-developer-cloud/discovery-components/compare/discovery-search-app@1.0.1-beta.29...discovery-search-app@1.1.0-beta.0) (2020-03-05)
+
+
+### Features
+
+* add DocumentPreview error message ([#104](https://github.com/watson-developer-cloud/discovery-components/issues/104)) ([e1a7442](https://github.com/watson-developer-cloud/discovery-components/commit/e1a7442))
+
+
+
+
+
+## [1.0.1-beta.29](https://github.com/watson-developer-cloud/discovery-components/compare/discovery-search-app@1.0.1-beta.28...discovery-search-app@1.0.1-beta.29) (2020-03-04)
+
+**Note:** Version bump only for package discovery-search-app
+
+
+
+
+
+## [1.0.1-beta.28](https://github.com/watson-developer-cloud/discovery-components/compare/discovery-search-app@1.0.1-beta.27...discovery-search-app@1.0.1-beta.28) (2020-03-03)
+
+**Note:** Version bump only for package discovery-search-app
+
+
+
+
+
+## [1.0.1-beta.27](https://github.com/watson-developer-cloud/discovery-components/compare/discovery-search-app@1.0.1-beta.26...discovery-search-app@1.0.1-beta.27) (2020-03-02)
+
+**Note:** Version bump only for package discovery-search-app
+
+
+
+
+
 ## [1.0.1-beta.26](https://github.com/watson-developer-cloud/discovery-components/compare/discovery-search-app@1.0.1-beta.25...discovery-search-app@1.0.1-beta.26) (2020-02-20)
 
 **Note:** Version bump only for package discovery-search-app